requirements.txt
PyPDF2==3.0.1
beautifulsoup4==4.12.3
python-pptx==0.6.23
python-docx==1.1.2
langdetect==1.0.9
python-magic-bin==0.4.14
pdfminer==20191125
readabilipy==0.2.0
fastapi==0.111.0
pandas
phonenumbers==8.13.38
intervaltree==3.1.0
base58==2.1.1
bech32==1.2.0
cashaddress==1.0.6
iteration_utilities==0.12.1
cbor==1.0.0
lxml==5.2.2
eth-hash==0.7.0